removed some of the redundent code
