--- CellFabric/tests/__json_default_fabric_beol_sticks_gold_old
@@ -1,282 +1,0 @@
-{
-  "bbox": [
-    4,
-    -20,
-    544,
-    320
-  ],
-  "globalRoutes": [],
-  "globalRouteGrid": [],
-  "terminals": [
-    {
-      "layer": "M1",
-      "netName": "a",
-      "rect": [
-        24,
-        6,
-        56,
-        258
-      ]
-    },
-    {
-      "layer": "M1",
-      "netName": "a",
-      "rect": [
-        184,
-        6,
-        216,
-        258
-      ]
-    },
-    {
-      "layer": "M1",
-      "netName": "a",
-      "rect": [
-        344,
-        6,
-        376,
-        258
-      ]
-    },
-    {
-      "layer": "M1",
-      "netName": "b",
-      "rect": [
-        104,
-        6,
-        136,
-        258
-      ]
-    },
-    {
-      "layer": "M1",
-      "netName": "b",
-      "rect": [
-        264,
-        6,
-        296,
-        258
-      ]
-    },
-    {
-      "layer": "M1",
-      "netName": "b",
-      "rect": [
-        424,
-        6,
-        456,
-        258
-      ]
-    },
-    {
-      "layer": "M2",
-      "netName": "b",
-      "rect": [
-        84,
-        194,
-        476,
-        226
-      ]
-    },
-    {
-      "layer": "M2",
-      "netName": "a",
-      "rect": [
-        4,
-        110,
-        396,
-        142
-      ]
-    },
-    {
-      "layer": "M2",
-      "netName": "b",
-      "rect": [
-        84,
-        26,
-        476,
-        58
-      ]
-    },
-    {
-      "layer": "V1",
-      "netName": "b",
-      "rect": [
-        104,
-        26,
-        136,
-        58
-      ]
-    },
-    {
-      "layer": "V1",
-      "netName": "b",
-      "rect": [
-        104,
-        194,
-        136,
-        226
-      ]
-    },
-    {
-      "layer": "V1",
-      "netName": "b",
-      "rect": [
-        264,
-        26,
-        296,
-        58
-      ]
-    },
-    {
-      "layer": "V1",
-      "netName": "b",
-      "rect": [
-        264,
-        194,
-        296,
-        226
-      ]
-    },
-    {
-      "layer": "V1",
-      "netName": "b",
-      "rect": [
-        424,
-        194,
-        456,
-        226
-      ]
-    },
-    {
-      "layer": "V1",
-      "netName": "a",
-      "rect": [
-        24,
-        110,
-        56,
-        142
-      ]
-    },
-    {
-      "layer": "V1",
-      "netName": "a",
-      "rect": [
-        184,
-        110,
-        216,
-        142
-      ]
-    },
-    {
-      "layer": "V1",
-      "netName": "a",
-      "rect": [
-        344,
-        110,
-        376,
-        142
-      ]
-    },
-    {
-      "layer": "M3",
-      "netName": "b",
-      "rect": [
-        420,
-        0,
-        460,
-        320
-      ]
-    },
-    {
-      "layer": "V2",
-      "netName": "b",
-      "rect": [
-        424,
-        26,
-        456,
-        58
-      ]
-    },
-    {
-      "layer": "V2",
-      "netName": "b",
-      "rect": [
-        424,
-        194,
-        456,
-        226
-      ]
-    },
-    {
-      "layer": "M4",
-      "netName": "b",
-      "rect": [
-        400,
-        180,
-        544,
-        220
-      ]
-    },
-    {
-      "layer": "M4",
-      "netName": "b",
-      "rect": [
-        400,
-        20,
-        544,
-        60
-      ]
-    },
-    {
-      "layer": "V3",
-      "netName": "b",
-      "rect": [
-        420,
-        20,
-        460,
-        60
-      ]
-    },
-    {
-      "layer": "V3",
-      "netName": "b",
-      "rect": [
-        420,
-        180,
-        460,
-        220
-      ]
-    },
-    {
-      "layer": "M5",
-      "netName": "b",
-      "rect": [
-        472,
-        -20,
-        536,
-        268
-      ]
-    },
-    {
-      "layer": "V4",
-      "netName": "b",
-      "rect": [
-        484,
-        20,
-        524,
-        60
-      ]
-    },
-    {
-      "layer": "V4",
-      "netName": "b",
-      "rect": [
-        484,
-        180,
-        524,
-        220
-      ]
-    }
-  ]
-}
